Parsing map file, collecting brush planes and brush parameters, and splitting planes into polygons. Need to make sure everything is in working order, then export the FBX with the brush polygons.
I ran into an issue with textures. I need the original texture sizes to be able to lay the texture onto the face and to generate the proper UVs. Since I don't have the texture sizes, I will have to recursively run through the games texture folders and accumulate the sizes at some point. So to be able to export proper UVs, the user will have to pass in a texture folder with the export call.